refactor(icons): move sizing utils
